refactor(Entire Application): Modularized entire app for greater Redux usage, separated out componen
Separated containers and components whereever possible for clearer code reading.
