+++ public/Styles/_eventList.scss
@@ -64,7 +64,8 @@
 .playlist-btn {
   background-color: $spotify-green;
   color: $white;
-  width: 100%;
+  width: 95%;
+  margin-left: 5%;
   margin-top: 20%;
 }
