+++ src/components/Carousel.jsx
@@ -1,0 +1,35 @@
+import React, { Component } from 'react';
+import Slider from 'react-slick';
+
+class Carousel extends Component {
+  render() {
+    const settings = {
+      dots: true,
+      dotsClass: 'slick-dots slick-thumb',
+      infinite: true,
+      fade: false,
+      speed: 1000,
+      slidesToShow: 1,
+      slidesToScroll: 1,
+      swipeToSlide: false,
+      autoplay: true,
+      autoplaySpeed: 6000,
+    };
+
+    return (
+      <div className="carousel">
+        <Slider {...settings}>
+          <div className="top-page-container">
+            <div className="home-carousel">
+              <img className="home-header-logo" src="./assets/gigify.svg"/>
+            </div>
+          </div>
+          <a className="carousel-image panorama" data-toggle="modal"  data-target="#loadingModal" onClick={() => this.handleFirst()} />
+          <a className="carousel-image govball" data-toggle="modal" data-target="#loadingModal" onClick={() => this.handleSecond()} />
+        </Slider>
+      </div>
+    )
+  }
+}
+
+export default Carousel;
