+++ src/constants/ActionTypes.js
@@ -1,5 +1,3 @@
-// export const DISPLAY_ENTRIES = 'DISPLAY_ENTRIES';
-// export const ADD_EVENT = 'ADD_EVENT';
 export const SELECT_EVENT = 'SELECT_EVENT';
 export const GET_EVENTS_DATA = 'GET_EVENTS_DATA';
 export const GET_EVENTS_DATA_RECEIVED = 'GET_EVENTS_DATA_RECEIVED';
@@ -8,4 +6,4 @@ export const SUBMIT_USERNAME = 'SUBMIT_USERNAME';
 export const SHOW_LOADING_PLAYLIST = 'SHOW_LOADING_PLAYLIST';
 export const SHOW_PLAYLIST = 'SHOW_PLAYLIST';
 export const PLAYLIST_ID = 'PLAYLIST_ID';
-// export const TOGGLE_ACTIVE = 'TOGGLE_ACTIVE';
+export const AUTHENTICATE_USER = 'AUTHENTICATE_USER';
