--- src/reducers/displayEvents.jsx
@@ -1,14 +1,0 @@
-// const initialState = {
-//   eventListings: [],
-// };
-//
-// const displayEvents = (state = initialState, action) => {
-//   switch (action.type) {
-//     case 'GET_EVENTS_DATA_RECEIVED':
-//       return { ...state, displayEvents: true };
-//     default:
-//       return state;
-//   }
-// };
-//
-// export default displayEvents;
