+++ src/reducers/index.js
@@ -1,12 +1,10 @@
 import { combineReducers } from 'redux';
 import events from './events';
-// import toggle from './toggle';
 import loading from './loading';
 import username from './username';
 import loadingplaylist from './loadingplaylist';
 import showplaylist from './showplaylist';
 import playlistid from './playlistid';
-// import displayEvents from './displayEvents'
 
 
 const eventsApp = combineReducers({
@@ -16,8 +14,6 @@ const eventsApp = combineReducers({
   loadingplaylist,
   showplaylist,
   playlistid,
-  // displayEvents,
-  // toggle,
 });
 
 export default eventsApp;
